requirements.txt
azure-functions
pandas
xmltodict
pyodbc
requests
cachetools==3.1.1
certifi==2020.4.5.1
chardet==3.0.4
click==6.7
Flask==1.0.2
google-api-core==1.17.0
google-auth==1.15.0
googleapis-common-protos==1.51.0
idna==2.9
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
protobuf==3.11.3
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
rsa==4.0
six==1.15.0
urllib3==1.25.9
Werkzeug==0.14.1